include: change tick's data type
Change tick data type to more generic type to lower the probability of
overflow.

+++ src/include/time.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-extern u32 tick;
+extern size_t tick;
 
 int arch_time_init(void);
 int time_init(void);
