+++ src/kernel/time.c
@@ -1,6 +1,6 @@
 #include <time.h>
 
-u32 tick = 0;
+size_t tick = 0;
 
 int time_init(void)
 {
